Forgot password done
Took forever but forgot password functionality is complete - there are probably a few checks I'm forgetting and there's still a lot to be done in terms of display and showing errors but whatever. I think the email service limits me to 1000 emails per day so let's hope that we don't hit that anytime soon.