Multi photo sort of
Multi photos are sort of working, buttons are going to be the easy part here. I have no clue how to size the images to fit the same size container... Might just keep them how they are and then move the "switch photo" buttons outside of the div so they don't go everywhere. They might fit in between the username of the poster and the date.