Basics of video uploads
Videos can now be uploaded by users - there is a maximum of 500MB per file uploaded. This may be lowered in the future... Uploading images taken on a phone result in a scroll bar too often, need to rework the image and video height for portrait media. Also, one video isn't showing but plays the audio and I have no idea why.