Added DB Messages
Messages are now committed to the DB upon sending to the websocket - more might be done on this, where a message is not committed to the DB until it comes back from the socket so as to confirm it made it. Messages are also fetched back from the DB upon page load and are displayed accordingly.