Login fixes
Ensured password reset was still working, in local env clicking the reset link from the email boots user directly back to login screen - this is because of the useEffect in the UserContext that wraps all components in the layout, but for some reason this does not affect users in production. Believe it may have something to do with Strict Mode?